pyproject.toml
[project]
name = "ehamtext"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.3",
    "docopt>=0.6.2",
    "fairseq2>=0.7.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "nltk>=3.9.3",
    "numpy>=1.26.4",
    "plotly>=6.6.0",
    "rich>=13.9.4",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sonar-space>=0.5.0",
    "torch>=2.9.0",
]
